+++ crates/test/src/config.rs
@@ -7,7 +7,7 @@ use std::path::Path;
 use serde::{Deserialize, Serialize};
 
 use smplx_regtest::RegtestConfig;
-use smplx_sdk::program::TrackerLogLevel;
+use smplx_sdk::global::Verbosity;
 
 use super::error::TestError;
 
@@ -22,7 +22,7 @@ pub struct TestConfig {
     pub bitcoins: u64,
     pub esplora: Option<EsploraConfig>,
     pub rpc: Option<RpcConfig>,
-    pub verbosity: Option<Verbosity>,
+    pub verbosity: Verbosity,
 }
 
 #[derive(Debug, Default, Clone, Serialize, Deserialize)]
@@ -36,23 +36,6 @@ pub struct RpcConfig {
     pub url: String,
     pub username: String,
     pub password: String,
-}
-
-#[derive(Debug, Clone, Copy, Serialize, Deserialize)]
-pub struct Verbosity(pub u64);
-
-impl TryInto<TrackerLogLevel> for Verbosity {
-    type Error = Self;
-
-    fn try_into(self) -> Result<TrackerLogLevel, Self::Error> {
-        match self {
-            Verbosity(1) => Ok(TrackerLogLevel::None),
-            Verbosity(2) => Ok(TrackerLogLevel::Warning),
-            Verbosity(3) => Ok(TrackerLogLevel::Debug),
-            Verbosity(4) => Ok(TrackerLogLevel::Trace),
-            _ => Err(self),
-        }
-    }
 }
 
 impl TestConfig {
@@ -97,7 +80,7 @@ impl Default for TestConfig {
             bitcoins: DEFAULT_BITCOINS,
             esplora: None,
             rpc: None,
-            verbosity: Some(Verbosity(3)),
+            verbosity: Verbosity::None,
         }
     }
 }